Sort files per set to avoid new inserts to be placed at the end
New files Injected during work were being placed at the end of the group without accounting for alphabetical order. However, when doing a fresh gulp run, the files would then sort alphabetically - producing a change in the index.html file and triggering a need to commit the file to repo. 
With this fix, the files are always sorts prior to injection.